+++ lib/usecases/posts/usecase_get_comments.dart
@@ -1,6 +1,6 @@
 import 'package:meta/meta.dart';
 import 'package:mooncake/entities/entities.dart';
-import 'package:mooncake/usecases/posts/posts.dart';
+import 'package:mooncake/usecases/usecases.dart';
 
 /// Allows to read the comments of a post having a specific id.
 class GetCommentsUseCase {
