+++ lib/usecases/posts/usecase_save_post.dart
@@ -1,6 +1,6 @@
 import 'package:meta/meta.dart';
 import 'package:mooncake/entities/entities.dart';
-import 'package:mooncake/usecases/posts/posts.dart';
+import 'package:mooncake/usecases/usecases.dart';
 
 /// Allows to save a new post into the repository.
 class SavePostUseCase {
